tests/wpt/web-platform-tests/speculation-rules/prerender/session-history-pushstate.https.html

<!DOCTYPE html>
<title>Test history.length</title>
<meta name="timeout" content="long">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/utils.js"></script>
<script src="../resources/utils.js"></script>
<script src="resources/utils.js"></script>
<script src="resources/session-history-test-util.js"></script>

<body>
<script>
setup(() => assertSpeculationRulesIsSupported());

const uid = token();

promise_test(async () => {
  assert_equals(
    await runTestInPrerender("testHistoryPushStateInPrerender", uid),
    "Passed"
  );
}, "history.pushState navigates independently with replacement in a prerender");

promise_test(async () => {
  assert_equals(
    await runTestInPrerender("testHistoryReplaceStateInPrerender", uid),
    "Passed"
  );
}, "history.replaceState navigates independently in a prerender");
</script>
